floating point issue solved

+++ demosales-ui/sales/src/main/java/com/imecatro/demosales/ui/sales/add/mappers/ProductDomainToUiMapper.kt
@@ -25,7 +25,7 @@ internal fun ProductDomainModel.toCartUiModel(): ProductOnCartUiModel {
     return ProductOnCartUiModel(
         product = this.toAddSaleUi(),
         qty = 0f,
-        subtotal = 0f
+        subtotal = 0f.toBigDecimal()
     )
 }
