+++ demosales-ui/sales/src/main/java/com/imecatro/demosales/ui/sales/add/model/ProductOnCartUiModel.kt
@@ -1,7 +1,9 @@
 package com.imecatro.demosales.ui.sales.add.model
 
+import java.math.BigDecimal
+
 data class ProductOnCartUiModel(
     val product: ProductResultUiModel,
     var qty: Float,
-    var subtotal: Float
+    var subtotal: BigDecimal
 )
